fashion/requirements.txt
langgraph
langchain
langchain-openai
openai
python-dotenv
